Added basic menu schema, remove unneeded includes
Cursor is now drawn when insert is held and is getting whole window input while disabling it for CSGO original fn.

First schema of the menu function has been made. Full TODO list has been made with everything it will include in the future. Don't know if I'll push these things to the pub. or just use a more simplistic design for this.