Added ComboBox to menu, some more styling fixes
Added global FontHeight member of each float to keep consistency as well as to use the least calculations to get the font height.

Added ComboBox to menu. This version of it is not looking pretty, but I'm not focusing on that, the look is quite easy to replace. Its code can and will be improved later on.